+++ mypy.ini
@@ -1,7 +1,7 @@
 [mypy]
 ignore_missing_imports = True
 
-[mypy-tabml.utils.pb_helpers]
+[mypy-tabml.utils.config_helpers]
 ignore_errors = True
 
 [mypy-tabml.protos.*]
