+++ tabml/feature_config_helper.py
@@ -1,7 +1,7 @@
 import copy
 from typing import Dict, List, Union
 
-from tabml.utils.pb_helpers import parse_feature_config_pb
+from tabml.utils.config_helpers import parse_feature_config_pb
 from tabml.utils.utils import check_uniqueness
