+++ tabml/model_wrappers.py
@@ -8,8 +8,8 @@ from xgboost import XGBClassifier, XGBRegressor
 
 from tabml.data_loaders import BaseDataLoader
 from tabml.utils import utils
+from tabml.utils.config_helpers import pb_to_dict
 from tabml.utils.logger import boosting_logger_eval
-from tabml.utils.pb_helpers import pb_to_dict
 from tabml.utils.utils import save_as_pickle
 
 MLFLOW_AUTOLOG = {
