+++ tests/test_data_loaders.py
@@ -6,7 +6,7 @@ import pytest
 
 from tabml.data_loaders import BaseDataLoader
 from tabml.feature_manager import BaseFeatureManager
-from tabml.utils.pb_helpers import parse_pipeline_config_pb
+from tabml.utils.config_helpers import parse_pipeline_config_pb
 from tabml.utils.utils import write_str_to_file
