+++ tests/test_feature_config_helper.py
@@ -2,7 +2,7 @@ import pytest
 from qcore.asserts import AssertRaises, assert_eq
 
 from tabml import feature_config_helper
-from tabml.utils.pb_helpers import parse_feature_config_pb
+from tabml.utils.config_helpers import parse_feature_config_pb
 from tabml.utils.utils import write_str_to_file
